Make assert its own directory
